Mapeamento type no postgrees

+++ src/ProjectName.Infrastructure.PostgreSQL/Repositories/OrderRepository.cs
@@ -17,7 +17,7 @@ public class OrderRepository: IOrderRepository
 
     private Order ToDomain(OrderEntity entity) => new Order(entity.Id,(OrderStatus)entity.Status,entity.TotalAmount);
 
-    private OrderEntity ToEntity(Order domain)=> new OrderEntity (domain.Id,(int)domain.Status, domain.TotalAmount);
+    private OrderEntity ToEntity(Order domain)=> new OrderEntity (domain.Id,domain.Status, domain.TotalAmount);
 
     public void Add(Order order)
     {
